requirements.txt
inflect==6.0.2
jsonschema==4.17.3
PyYAML==6.0
networkx==2.8.8
numpy==1.23.5
pandas==1.5.2
keras==2.11.0
h5py==3.7.0
matplotlib==3.6.2
matplotlib-inline==0.1.6
pygraphviz==1.10
pydot==1.4.2
pytz==2022.6
seaborn==0.12.1
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.11.0
Wand==0.6.11
texoutparse==1.0.0
tex2pix==0.3.1